Adding common GitHub Event Class to eliminate duplicate code
