+++ src/main/java/com/checkmarx/flow/dto/github/EventCommon.java
@@ -1,4 +1,3 @@
-
 package com.checkmarx.flow.dto.github;
 
 import com.fasterxml.jackson.annotation.*;
@@ -8,50 +7,15 @@ import java.util.Map;
 
 @JsonIgnoreProperties
 @JsonInclude(JsonInclude.Include.NON_NULL)
-public class PingEvent extends EventCommon {
-
-    @JsonProperty("zen")
-    private String zen;
-    @JsonProperty("hook_id")
-    private Integer hookId;
-    @JsonProperty("hook")
-    private Hook hook;
+public class EventCommon {
     @JsonProperty("repository")
     private Repository repository;
     @JsonProperty("sender")
     private Sender sender;
+    @JsonProperty("installation")
+    private Installation installation;
     @JsonIgnore
     private Map<String, Object> additionalProperties = new HashMap<>();
-
-    @JsonProperty("zen")
-    public String getZen() {
-        return zen;
-    }
-
-    @JsonProperty("zen")
-    public void setZen(String zen) {
-        this.zen = zen;
-    }
-
-    @JsonProperty("hook_id")
-    public Integer getHookId() {
-        return hookId;
-    }
-
-    @JsonProperty("hook_id")
-    public void setHookId(Integer hookId) {
-        this.hookId = hookId;
-    }
-
-    @JsonProperty("hook")
-    public Hook getHook() {
-        return hook;
-    }
-
-    @JsonProperty("hook")
-    public void setHook(Hook hook) {
-        this.hook = hook;
-    }
 
     @JsonProperty("repository")
     public Repository getRepository() {
@@ -73,6 +37,16 @@ public class PingEvent extends EventCommon {
         this.sender = sender;
     }
 
+    @JsonProperty("installation")
+    public Installation getInstallation() {
+        return installation;
+    }
+
+    @JsonProperty("installation")
+    public void setInstallation(Installation installation) {
+        this.installation = installation;
+    }
+
     @JsonAnyGetter
     public Map<String, Object> getAdditionalProperties() {
         return this.additionalProperties;
@@ -82,5 +56,4 @@ public class PingEvent extends EventCommon {
     public void setAdditionalProperty(String name, Object value) {
         this.additionalProperties.put(name, value);
     }
-
 }
